RulesParser.RulesParser: constructor argument enhancements
The RulesParser constructor code (largely cut-pasted from ply examples)
does a piss-poor job of respecting the passwd in kwargs.  Clearly, the
sane thing to do is to respect as many arguments as are applicable to
both the lexer and the parser.  To ensure that we do so, and to avoid
creating argument conflicts between kwargs and the hard-coded arguments,
we create a separate whitelist of non-special-cased arguments for lex.lex and
yacc.yacc invocations.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>