RulesParser.RulesParser: outermost parsing rules
Rewire the outermost parsing rules, a bit, to something that might
actually stand a chance of working out:

add an outermost "rulesprogram" production.  This may come in handy
later if there is need for some special constructs in the global
context, or, on the production side, if there is some final optimization
phase or other filtering that should be applied once the entire
parse-tree is finalized for a program (although these could also be
applied externally, given that "we are done parsing" and would otherwise
just return our result).

Change the "statements" production to use left-binding-friendly
grammar rule.

Move the semicolon slurping into the statement production.  This will
eventually turn into a disjunction of possible statement types, so
it will wind up looking like:

  '''statement : assignment SEMICOLON
               | conditional SEMICOLON
	       | functioncall SEMICOLON
	       .
	       .
	       .'''

If I ever want to support c-style context-dependant semicolon skipping,
I'll have to rethink this entirely, probably putting it back into
statements.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>