RulesParser: implement lextab handling
This adds handling of lextab that is roughly analogous to what is
already happening with parsetab, both at the NewLexer level and the
Luthor level of inheritance.

One major differece is that, when a lextab module is found within a
package, we have to set optimize=1, if it is to be used.  We only do
this automatically in the case where -OO optimizations are active in
python and therefore we are going to fail unless we resort to it, since
this could make debugging quite difficult otherwise.

Signed-off-by: Gregory M. Turner <gmt@be-evil.net>